Move useCustomUnits
