+++ packages/block-editor/src/components/index.js
@@ -82,6 +82,7 @@ export { default as URLInputButton } from './url-input/button';
 export { default as URLPopover } from './url-popover';
 export { __experimentalImageURLInputUI } from './url-popover/image-url-input-ui';
 export { default as withColorContext } from './color-palette/with-color-context';
+export { useCustomUnits } from './use-custom-units';
 
 /*
  * Content Related Components
