+++ packages/block-editor/src/hooks/padding.js
@@ -12,7 +12,7 @@ import { __experimentalBoxControl as BoxControl } from '@wordpress/components';
 import useSetting from '../components/use-setting';
 import { SPACING_SUPPORT_KEY, useCustomSides } from './spacing';
 import { cleanEmptyObject } from './utils';
-import { useCustomUnits } from '../components/unit-control';
+import { useCustomUnits } from '../components/use-custom-units';
 
 const isWeb = Platform.OS === 'web';
 const CSS_UNITS = [
